move screenshot templates to utils
